refactor: removed few comments and minor changes

+++ cli/commands/init.ts
@@ -1,17 +1,19 @@
 import { resolve } from 'path'
-import { promises } from 'fs'
+import { existsSync, promises } from 'fs'
 
-const config = `
-  module.exports = {
-    content: [],
-  }
+const config = `module.exports = {
+  content: [],
+}
 `
 
 export async function init() {
   const root = process.cwd()
 
-  await promises.writeFile(
-    resolve(root, 'awast.config.js'),
-    Buffer.from(config, 'utf-8')
-  )
+  if (!existsSync('./awast.config.js')) {
+    await promises.writeFile(
+      resolve(root, 'awast.config.js'),
+      Buffer.from(config, 'utf-8')
+    )
+    console.log('created awast.config.js')
+  }
 }
